[pkg/stanza] Deduplicate file rotation tests
These tests were copied from the file_input operator into the fileconsumer package
at the time the latter was extracted from the former. At this point it is clear
that the fileconsumer package is fully responsible for the behavior being tested,
so removing the test from file_input makes sense.
